requirements.txt
faiss-cpu
sentence-transformers
PyMuPDF
pytesseract
pdf2image
tqdm
google-generativeai
Pillow
python-docx
easyocr

